Factor out CoreReading component
The Core Reading block (with R:A-Z, The Codex, and HPMoR) was
copy-pasted into two places. Fix that.
