+++ packages/lesswrong/lib/components.js
@@ -246,6 +246,7 @@ import '../components/sequences/CollectionsEditForm.jsx';
 import '../components/sequences/BooksNewForm.jsx';
 import '../components/sequences/BooksEditForm.jsx';
 import '../components/sequences/BooksItem.jsx';
+import '../components/sequences/CoreReading.jsx';
 import '../components/collections/CollectionsCardContainer.jsx';
 import '../components/collections/CollectionsCard.jsx';
 import '../components/collections/BigCollectionsCard.jsx';
